Dev improve check permission

+++ PlayZone.API/Controllers/Worktime_Related/WorktimeController.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using PlayZone.API.Attributes;
@@ -21,15 +22,32 @@ public class WorktimeController : ControllerBase
         this._worktimeService = worktimeService;
     }
 
+    public static bool CheckHasPermissionWorktime(ClaimsPrincipal user, int id)
+    {
+        if (user.HasClaim("Permission", Permission.PERSO_CONSULTER_POINTAGE) &&
+            !user.HasClaim("Permission", Permission.ALL_CONSULTER_POINTAGES) &&
+            user.HasClaim(ClaimTypes.NameIdentifier, id.ToString()) == false)
+        {
+            return false;
+        }
+        return true;
+    }
+
     [HttpGet("range")]
-    [PermissionAuthorize(Permission.DEBUG_PERMISSION)]
+    [Authorize]
+    [PermissionAuthorize([Permission.PERSO_CONSULTER_POINTAGE, Permission.ALL_CONSULTER_POINTAGES])]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<WorktimeDTO>))]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public IActionResult GetByDateRange([FromQuery] int userId, [FromQuery] DateTime startDate, [FromQuery] DateTime endDate)
     {
+        if (!CheckHasPermissionWorktime(this.User, userId))
+            return this.StatusCode(StatusCodes.Status403Forbidden);
+
         try
         {
-            IEnumerable<WorktimeDTO> worktimes = this._worktimeService.GetByDateRange(userId, startDate, endDate).Select(w => w.ToDTO());
+            IEnumerable<WorktimeDTO> worktimes =
+                this._worktimeService.GetByDateRange(userId, startDate, endDate).Select(w => w.ToDTO());
             return this.Ok(worktimes);
         }
         catch (Exception)
@@ -39,13 +57,20 @@ public class WorktimeController : ControllerBase
     }
 
     [HttpGet("day/{userId:int}/{dayOfMonth:int}/{monthOfYear:int}/{year:int}")]
-    [PermissionAuthorize(Permission.DEBUG_PERMISSION)]
+    [Authorize]
+    [PermissionAuthorize([Permission.PERSO_CONSULTER_POINTAGE, Permission.ALL_CONSULTER_POINTAGES])]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<WorktimeDTO>))]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public IActionResult GetByDay(int userId, int dayOfMonth, int monthOfYear, int year)
     {
+        if (!CheckHasPermissionWorktime(this.User, userId))
+            return this.StatusCode(StatusCodes.Status403Forbidden);
+
         try
         {
-            IEnumerable<WorktimeDTO> worktimes = this._worktimeService.GetByDay(userId, dayOfMonth, monthOfYear, year).Select(w => w.ToDTO());
+            IEnumerable<WorktimeDTO> worktimes = this._worktimeService.GetByDay(userId, dayOfMonth, monthOfYear, year)
+                .Select(w => w.ToDTO());
             return this.Ok(worktimes);
         }
         catch (Exception)
@@ -55,13 +80,20 @@ public class WorktimeController : ControllerBase
     }
 
     [HttpGet("week/{userId:int}/{weekOfYear:int}/{year:int}")]
-    [PermissionAuthorize(Permission.DEBUG_PERMISSION)]
+    [Authorize]
+    [PermissionAuthorize([Permission.PERSO_CONSULTER_POINTAGE, Permission.ALL_CONSULTER_POINTAGES])]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<WorktimeDTO>))]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public IActionResult GetByWeek(int userId, int weekOfYear, int year)
     {
+        if (!CheckHasPermissionWorktime(this.User, userId))
+            return this.StatusCode(StatusCodes.Status403Forbidden);
+
         try
         {
-            IEnumerable<WorktimeDTO> worktimes = this._worktimeService.GetByWeek(userId, weekOfYear, year).Select(w => w.ToDTO());
+            IEnumerable<WorktimeDTO> worktimes =
+                this._worktimeService.GetByWeek(userId, weekOfYear, year).Select(w => w.ToDTO());
             return this.Ok(worktimes);
         }
         catch (Exception)
@@ -71,13 +103,20 @@ public class WorktimeController : ControllerBase
     }
 
     [HttpGet("month/{userId:int}/{monthOfYear:int}/{year:int}")]
-    [PermissionAuthorize(Permission.DEBUG_PERMISSION)]
+    [Authorize]
+    [PermissionAuthorize([Permission.PERSO_CONSULTER_POINTAGE, Permission.ALL_CONSULTER_POINTAGES])]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<WorktimeDTO>))]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public IActionResult GetByMonth(int userId, int monthOfYear, int year)
     {
+        if (!CheckHasPermissionWorktime(this.User, userId))
+            return this.StatusCode(StatusCodes.Status403Forbidden);
+
         try
         {
-            IEnumerable<WorktimeDTO> worktimes = this._worktimeService.GetByMonth(userId, monthOfYear, year).Select(w => w.ToDTO());
+            IEnumerable<WorktimeDTO> worktimes =
+                this._worktimeService.GetByMonth(userId, monthOfYear, year).Select(w => w.ToDTO());
             return this.Ok(worktimes);
         }
         catch (Exception)
@@ -87,8 +126,10 @@ public class WorktimeController : ControllerBase
     }
 
     [HttpGet("{id:int}")]
-    [PermissionAuthorize(Permission.DEBUG_PERMISSION)]
+    [Authorize]
+    [PermissionAuthorize([Permission.PERSO_CONSULTER_POINTAGE, Permission.ALL_CONSULTER_POINTAGES])]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(WorktimeDTO))]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(string))]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public IActionResult GetById(int id)
@@ -97,7 +138,12 @@ public class WorktimeController : ControllerBase
         {
             WorktimeDTO? worktime = this._worktimeService.GetById(id)?.ToDTO();
             if (worktime != null)
+            {
+                if (!CheckHasPermissionWorktime(this.User, worktime.UserId))
+                    return this.StatusCode(StatusCodes.Status403Forbidden);
                 return this.Ok(worktime);
+            }
+
             return this.NotFound("La plage horaire est introuvable.");
         }
         catch (Exception)
@@ -108,12 +154,16 @@ public class WorktimeController : ControllerBase
 
     [HttpPost]
     [Authorize]
-    [PermissionAuthorize(Permission.DEBUG_PERMISSION)]
+    [PermissionAuthorize([Permission.PERSO_CONSULTER_POINTAGE, Permission.ALL_CONSULTER_POINTAGES])]
     [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(WorktimeDTO))]
     [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(string))]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public IActionResult Create([FromBody] WorktimeUpdateFormDTO worktime)
     {
+        if (!CheckHasPermissionWorktime(this.User, worktime.UserId))
+            return this.StatusCode(StatusCodes.Status403Forbidden);
+
         try
         {
             int resultId = this._worktimeService.Create(worktime.ToModel());
@@ -127,17 +177,26 @@ public class WorktimeController : ControllerBase
         {
             return this.BadRequest(e.Message);
         }
-        catch (Exception) { /* ignored */ }
+        catch (Exception)
+        {
+            /* ignored */
+        }
 
         return this.StatusCode(StatusCodes.Status500InternalServerError);
     }
 
     [HttpPut("{id}")]
-    [PermissionAuthorize(Permission.DEBUG_PERMISSION)]
+    [Authorize]
+    [PermissionAuthorize([Permission.PERSO_CONSULTER_POINTAGE, Permission.ALL_CONSULTER_POINTAGES])]
     [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public IActionResult Update(int id, [FromBody] WorktimeUpdateFormDTO worktime)
     {
+        if (!CheckHasPermissionWorktime(this.User, worktime.UserId))
+            return this.StatusCode(StatusCodes.Status403Forbidden);
+
         if (id <= 0)
         {
             return this.BadRequest("Invalid user data");
@@ -149,15 +208,27 @@ public class WorktimeController : ControllerBase
         {
             return this.Ok();
         }
+
         return this.StatusCode(StatusCodes.Status500InternalServerError);
     }
 
     [HttpDelete("{idWorktime:int}")]
-    [PermissionAuthorize(Permission.DEBUG_PERMISSION)]
+    [Authorize]
+    [PermissionAuthorize([Permission.PERSO_CONSULTER_POINTAGE, Permission.ALL_CONSULTER_POINTAGES])]
     [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public IActionResult Delete(int idWorktime)
     {
+        WorktimeDTO? worktime = this._worktimeService.GetById(idWorktime)?.ToDTO();
+
+        if (worktime == null)
+            return this.StatusCode(StatusCodes.Status404NotFound);
+
+        if (!CheckHasPermissionWorktime(this.User, worktime.UserId))
+            return this.StatusCode(StatusCodes.Status403Forbidden);
+
         try
         {
             return this.Ok(this._worktimeService.Delete(idWorktime));
